Extract faces.ajax into faces/ajax.ts module #5506
